BUG=104111
TEST='QuotaManagerTest.*'


Review URL: http://codereview.chromium.org/8528022

TBR=tzik@chromium.org
Review URL: http://codereview.chromium.org/8572029

git-svn-id: svn://svn.chromium.org/chrome/branches/912/src@110183 0039d316-1c4b-4281-b951-d872f2087c98

+++ webkit/quota/quota_manager.cc
@@ -661,7 +661,7 @@ class QuotaManager::UpdatePersistentHostQuotaTask
   UpdatePersistentHostQuotaTask(
       QuotaManager* manager,
       const std::string& host,
-      int new_quota,
+      int64 new_quota,
       HostQuotaCallback* callback)
       : DatabaseTaskBase(manager),
         host_(host),
